tests/wpt/tests/navigation-api/navigation-methods/navigate-intercept-history-state.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async () => {
  history.replaceState("state1", "", "#1");
  assert_equals(history.state, "state1");

  navigation.onnavigate = e => e.intercept({ handler: () => Promise.resolve("r") });
  await navigation.navigate("#2").committed;
  assert_equals(location.hash, "#2");
  assert_equals(history.state, null);
}, "history.story should be nulled by navigate() handled by intercept()");
</script>
