LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/navigate-intercept-history-state-expected.txt


FAIL history.story should be nulled by navigate() handled by intercept() assert_equals: expected "#2" but got "#1"
